Code refactor with useSwitchGoogleAccount.
