+++ js/src/components/google-account-card/connected-google-account-card.js
@@ -6,55 +6,14 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { glaData } from '.~/constants';
-import { API_NAMESPACE } from '.~/data/constants';
 import AccountCard, { APPEARANCE } from '.~/components/account-card';
 import AppButton from '.~/components/app-button';
 import ConnectedIconLabel from '.~/components/connected-icon-label';
-import useGoogleAuthorization from '.~/hooks/useGoogleAuthorization';
-import useDispatchCoreNotices from '.~/hooks/useDispatchCoreNotices';
 import Section from '.~/wcdl/section';
-import useApiFetchCallback from '.~/hooks/useApiFetchCallback';
+import useSwitchGoogleAccount from './useSwitchGoogleAccount';
 
 export default function ConnectedGoogleAccountCard( { googleAccount } ) {
-	const pageName = glaData.mcSetupComplete ? 'reconnect' : 'setup-mc';
-	const { createNotice, removeNotice } = useDispatchCoreNotices();
-	const [
-		fetchGoogleDisconnect,
-		{ loading: loadingGoogleDisconnect },
-	] = useApiFetchCallback( {
-		method: 'DELETE',
-		path: `${ API_NAMESPACE }/google/connect`,
-	} );
-	const [
-		fetchGoogleConnect,
-		{ loading: loadingGoogleConnect, data: dataGoogleConnect },
-	] = useGoogleAuthorization( pageName );
-
-	const handleSwitch = async () => {
-		const { notice } = await createNotice(
-			'info',
-			__(
-				'Disconnecting your current Google account, please wait…',
-				'google-listings-and-ads'
-			)
-		);
-
-		try {
-			await fetchGoogleDisconnect();
-			const { url } = await fetchGoogleConnect();
-			window.location.href = url;
-		} catch ( error ) {
-			removeNotice( notice.id );
-			createNotice(
-				'error',
-				__(
-					'Unable to connect to a different Google account. Please try again later.',
-					'google-listings-and-ads'
-				)
-			);
-		}
-	};
+	const [ handleSwitch, loading ] = useSwitchGoogleAccount();
 
 	return (
 		<AccountCard
@@ -65,11 +24,7 @@ export default function ConnectedGoogleAccountCard( { googleAccount } ) {
 			<Section.Card.Footer>
 				<AppButton
 					isLink
-					disabled={
-						loadingGoogleDisconnect ||
-						loadingGoogleConnect ||
-						dataGoogleConnect
-					}
+					disabled={ loading }
 					text={ __(
 						'Or, connect to a different Google account',
 						'google-listings-and-ads'
